turn reference backlinks into protocol, make EditPreview conform to it
