+++ Wikipedia/Code/ArticleViewController+References.swift
@@ -73,36 +73,9 @@ extension ArticleViewController {
             popoverVC.scrollEnabled = true
         }
     }
-    
-    func dismissReferencesPopover() {
-        guard presentedViewController is WMFReferencePopoverMessageViewController || presentedViewController is WMFReferencePageViewController else {
-            return
-        }
-        dismiss(animated: true)
-    }
-    
+
     @objc func tappedWebViewBackground() {
         dismissReferenceBackLinksViewController()
-    }
-    
-    func showReferenceBackLinks(_ backLinks: [ReferenceBackLink], referenceId: String, referenceText: String) {
-        guard let vc = ReferenceBackLinksViewController(referenceId: referenceId, referenceText: referenceText, backLinks: backLinks, delegate: self, theme: theme) else {
-            showGenericError()
-            return
-        }
-        addChild(vc)
-        view.wmf_addSubviewWithConstraintsToEdges(vc.view)
-        vc.didMove(toParent: self)
-        referenceWebViewBackgroundTapGestureRecognizer.isEnabled = true
-    }
-    
-    func dismissReferenceBackLinksViewController() {
-        let vc = children.first { $0 is ReferenceBackLinksViewController }
-        vc?.willMove(toParent: nil)
-        vc?.view.removeFromSuperview()
-        vc?.removeFromParent()
-        messagingController.removeElementHighlights()
-        referenceWebViewBackgroundTapGestureRecognizer.isEnabled = false
     }
 }
 
@@ -159,7 +132,7 @@ extension ArticleViewController: UIPageViewControllerDelegate {
     }
 }
 
-extension ArticleViewController: WMFReferencePageViewAppearanceDelegate {
+extension ArticleViewController: WMFReferencePageViewAppearanceDelegate, ReferenceBackLinksViewControllerDelegate {
     func referencePageViewControllerWillAppear(_ referencePageViewController: WMFReferencePageViewController) {
         guard
             let firstRefVC = referencePageViewController.pageViewController.viewControllers?.first as? WMFReferencePanelViewController,
@@ -176,40 +149,8 @@ extension ArticleViewController: WMFReferencePageViewAppearanceDelegate {
     }
 }
 
-
-extension ArticleViewController: ReferenceBackLinksViewControllerDelegate {
-    func referenceViewControllerUserDidTapClose(_ vc: ReferenceViewController) {
-        if vc is ReferenceBackLinksViewController {
-            dismissReferenceBackLinksViewController()
-        } else {
-            dismissReferencesPopover()
-        }
-    }
-    
-    func referenceBackLinksViewControllerUserDidNavigateTo(referenceBackLink: ReferenceBackLink, referenceBackLinksViewController: ReferenceBackLinksViewController) {
-        scroll(to: referenceBackLink.id, centered: true, highlighted: true, animated: true)
-    }
-    
-    func referenceViewControllerUserDidNavigateBackToReference(_ vc: ReferenceViewController) {
-        referenceViewControllerUserDidTapClose(vc)
-        guard let referenceId = vc.referenceId else {
-            showGenericError()
-            return
-        }
-        scroll(to: "back_link_\(referenceId)", highlighted: true, animated: true) { [weak self] in
-            dispatchOnMainQueueAfterDelayInSeconds(1.0) { [weak self] in
-                self?.messagingController.removeElementHighlights()
-            }
-        }
-    }
-}
-
 extension ArticleViewController: UIGestureRecognizerDelegate {
     func gestureRecognizer(_ gestureRecognizer: UIGestureRecognizer, shouldRecognizeSimultaneouslyWith otherGestureRecognizer: UIGestureRecognizer) -> Bool {
-        if gestureRecognizer === referenceWebViewBackgroundTapGestureRecognizer {
-            return true
-        }
-        
-        return false //default
+        return shouldRecognizeSimultaneousGesture(recognizer: gestureRecognizer)
     }
 }
